# Conflicts:
#	src/app/app.module.ts

+++ src/app/additional-literature/additional-literature.component.ts
@@ -3,6 +3,7 @@ import {AdditionalLiteratureService} from '../services/additional-literature.ser
 import {AdditionalLiterature} from '../models/AdditionalLiterature';
 import {MDBModalRef, MDBModalService, MdbTableDirective, MdbTablePaginationComponent} from 'angular-bootstrap-md';
 import {AdditionalLiteratureEditComponent} from './additional-literature-edit.component';
+import {IsDeleteComponent} from '../is-delete/is-delete.component';
 
 @Component({
   selector: 'app-additional-literature',
@@ -98,8 +99,14 @@ export class AdditionalLiteratureComponent implements OnInit, AfterViewInit {
   }
 
   delete(el: any): void {
-    this.valueService.deleteValue(el.first).subscribe(() => {
-      this.removeRow(el);
+    const editableRow = {heading: el.last};
+    this.modalRef = this.modalService.show(IsDeleteComponent, this.modalOption(editableRow));
+    this.modalRef.content.saveButtonClicked.subscribe((newElement: any) => {
+      if (newElement) {
+        this.valueService.deleteValue(el.first).subscribe(() => {
+          this.removeRow(el);
+        });
+      }
     });
   }
